Fix parser and unit tests

+++ character.cpp
@@ -146,7 +146,7 @@ Character* Character::parseUnit(const std::string charSheetName)
 	}
 
 	if(attributes.find("name")!=attributes.end() && attributes.find("health")!=attributes.end() && attributes.find("dmg")!=attributes.end()){
-			return new Character(attributes["name"], std::stoi(attributes["health"]), std::stoi(attributes["dmg"]), std::stoi(attributes["atc"]));
+			return new Character(attributes["name"], std::stoi(attributes["health"]), std::stoi(attributes["dmg"]), std::stod(attributes["atc"]));
 	}
 	else{
 			throw "Invalid attributes in " + charSheetName + '\n';
